livekit-interrupt-handler-yash/requirements.txt
livekit-agents[openai, cartesia, deepgram, silero, turn-detector, mcp]>=1.0
python-dotenv
pytest
